ostree: Fix boot regex for directory
Issue: LINCD-9344

The @OSTREE_OSNAME@ hasn't been replaced, so there is an error when run
/usr/lib/systemd/system-generators/ostree-system-generator:

systemd[252]: /usr/lib/systemd/system-generators/ostree-system-generator failed with exit status 1.

And replace it in the source isn't a good way since it changes the source, so
use "/[^/]+" to replace @OSTREE_OSNAME@ to fix the problem.

(LOCAL REV: NOT UPSTREAM) -- Will send to meta-lat with other commits later

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>