conf: efitools: avoid dangling bbappend
Issue: LINUXEXEC-23251

If we run yocto-check-layer or do a build with meta-lat present but
meta-efi-secure-boot not available we will get a dangling bbappend
ERROR. Such as

    yocto-check-layer-wrapper ../layers/wr-ota --dependency \
      ../layers/meta-openembedded ../layers/meta-virtualization \
      ../layers/meta-lat
    [...]
    ERROR: No recipes in default available for:
      .../layers/meta-lat/recipes-bsp/efitools/efitools_%.bbappend

We could solve this simply by adding 'efi-secure-boot' to LAYERDEPENDS
and be done with it. But what is described in efitools bbappend along
with other references to the meta-efi-secure-boot layer are all
dependent on the DISTRO_FEATURE 'efi-secure-boot', so we can respect
this.

We make 'efi-secure-boot' in LAYERDEPENDS dependent on the
'efi-secure-boot' DISTRO_FEATURE, such that if it is present an ERROR
will be shown for the missing dependency. This ensures the layer is
available when it will in fact be used.

We make the bbappends only applicable if the meta-efi-secure-boot
layer is present, this will ensure it is never dangling. The automatic
adjustment to LAYERDEPENDS will ensure it is used when needed,
ie. when 'efi-secure-boot' (and 'ostree') DISTRO_FEATURE are present.

(master rev: bf94302f11a94236658e2ee31a27fe064b65a1f9)

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Mingli Yu <mingli.yu@windriver.com>