lat: correct memory region for hypervisor image
The region of memory 0xa0000000-0xa03fffff is reserved for communicating with the
RDC via an SCMI call. Unfortunately, our FIT configuration is placing the VxWorks
guest starting at 0xa0000000, and this results in the lower region of RAM for
VxWorks being corrupted by the SCMI call.

When corruption occurs the VxWorks kernel's heap becomes corrupted resulting in
bad pointers being followed on freeing memory.  This typically shows up when a
physical device is assigned to VxWorks.

Signed-off-by: Mark Dapoz <Mark.Dapoz@windriver.com>

Issue: LINUXEXEC-23467
(LOCAL REV: NOT UPSTREAM) -- Will send to meta-lat with other commits later

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>