meta-lat: remove setting WRL_RECIPES from template
Issue: LINUXEXEC-26075

Some customers want access to use recipes that are included in
WRL_RECIPES but they do not use templates. By moving the updates to
WRL_RECIPES outside of the feature templates we can simplify this
usecase.

We move the changes to WRL_RECIPES to wrlinux-recipelists.conf.

(LOCAL REV: NOT UPSTREAM) -- WRLinux specific

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>